requirements.txt
openai-whisper
youtube-transcript-api
yt-dlp
selenium
torch
